Crate snake_case

Structs

InvalidSnakeCase
Only one possible error: the given string was not valid snake_case.

SnakeCase
An owning string type that can only contain valid snake_case. In other words, it always matches ^[_a-z][_a-z0-9]*$

SnakeCaseRef
An non-owning string type that can only refer to string containing valid snake_case. In other words, it always matches ^[_a-z][_a-z0-9]*$

Functions

is_snake_case
Is the given string a non-empty snake_case string? In particular, does it match ^[_a-z][_a-z0-9]*$ ?